House Price Prediction
There are various factors which are considered while purchasing a house. To determine the best price of the house it is cruicial to evaluate each variable and understand the role it is playing in increasing or decreasing the price of the house. In this dataset, we have 81 variables which are evaluated to create a model that can accurately predict the price of the house.